+++ packages/cli-plugin-test/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.7-beta.1](https://github.com/midwayjs/cli/compare/v1.3.6...v1.3.7-beta.1) (2022-07-06)
+
+
+### Bug Fixes
+
+* common http daya type ([3498985](https://github.com/midwayjs/cli/commit/3498985c10fc21b47dbacf07b44068bb82c6b815))
+* mocha reporter ([d886999](https://github.com/midwayjs/cli/commit/d8869996c54a52efb0774a4a11f91489a4a8aaec))
+* **test:** avoid passing ignoreTypeCheck to jest ([#276](https://github.com/midwayjs/cli/issues/276)) ([7e76397](https://github.com/midwayjs/cli/commit/7e763979c80d0e4f2cfdf36bd7e7a074f5cd10a7))
+
+
+
+
+
 ## [1.3.5](https://github.com/midwayjs/cli/compare/v1.3.5-beta.3...v1.3.5) (2022-05-25)
 
 **Note:** Version bump only for package @midwayjs/cli-plugin-test